Create EPIC-VM-Home
For this script to work the VM tag must include the home ESXi host name, I used “-Like” ESXi name, so FQDN should not be required for it to work. It also doesn’t prompt for credentials, it just passes through.
The script is set to send out an email, but you can remove 2x “#” in lines 26 and 37 to output the data on a local powershell session.
If a tag is not set for a VM in the defined cluster, it will email/printout a tag needs to be set.